+++ aws_codeseeder/resources/npm_mirror_support.py
@@ -29,8 +29,11 @@ def get_secret(secret_name: str) -> Dict[str, Dict[str, str]]:
 
 
 def main(url: str) -> None:
-    secret_name = os.environ.get("AWS_CODESEEDER_MIRROR_SECRET")
-    if secret_name:
+    secret_name = os.environ.get("AWS_CODESEEDER_NPM_MIRROR_SECRET", "NO_SECRET")
+    # Backwards Compatibility
+    if secret_name == "NO_SECRET":
+        secret_name = os.environ.get("AWS_CODESEEDER_MIRROR_SECRET", "NO_SECRET")
+    if secret_name != "NO_SECRET":
         secret_name_key = secret_name.split("::")[0] if "::" in secret_name else secret_name
         key = secret_name.split("::")[1] if "::" in secret_name else "npm"
         creds = get_secret(secret_name=secret_name_key)
@@ -39,7 +42,7 @@ def main(url: str) -> None:
         logger.info("Calling npm config with %s", url)
         subprocess.call(["npm", "config", "set", f"{url.replace('https:', '')}:_auth={ssl_token}"])
     else:
-        logger.info("'AWS_CODESEEDER_MIRROR_SECRET' is not set")
+        logger.info("'AWS_CODESEEDER_NPM_MIRROR_SECRET' is not set")
 
 
 if __name__ == "__main__":