+++ CHANGELOG.md
@@ -9,7 +9,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ---
 
 ### New
-  
+- adds support for npm mirrors to be set
+
 ### Changes
 
 ### Fixes