+++ CHANGELOG.md
@@ -9,7 +9,13 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ---
 
 ### New
-  
+- adds support for npm mirrors to be set
+  - the SecretsManager secret name MUST be of the form `*-mirror-credentials`
+  - the content must be JSON
+  - the content has the form `{
+      "npm": { "ssl_token": "base64encodedtoken"},
+    }`
+
 ### Changes
 
 ### Fixes