+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
       "npm": { "ssl_token": "base64encodedtoken"},
     }`
   - npm registry is set via `npmMirror`
+    - `ssl_token` and `npmMirrorSecret` are *optional* when using a public mirror.
 ### Changes
 - Add support for explicit secret environment variables for mirrors